Add TextArea option for C11 and fix some build inconsistencies
