+++ platformio.ini
@@ -37,6 +37,14 @@
 build_flags =
   -Wall -Wno-parentheses -fdiagnostics-show-option
 
+[minimal_size]
+build_flags = 
+  -Os
+  -ffunction-sections 
+  -fdata-sections
+  -Wl,--gc-sections
+  -s
+
 
 [espota]
 upload_protocol = espota
@@ -91,8 +99,10 @@ build_flags               = -D BUILD_GIT='"${sysenv.TRAVIS_TAG}"'
 [esp82xx_2_5_x]
 build_flags               = -D BUILD_GIT='"${sysenv.TRAVIS_TAG}"'
                             -DNDEBUG
+                            -mtarget-align
                             -DVTABLES_IN_FLASH
                             -fno-exceptions
+                            -lstdc++
                             -DPIO_FRAMEWORK_ARDUINO_LWIP2_LOW_MEMORY_LOW_FLASH
                             -DWEBSERVER_RULES_DEBUG=0
                             -DPUYA_SUPPORT=1
@@ -196,7 +206,7 @@ build_flags               = -DMQTT_MAX_PACKET_SIZE=1024
 
 [common]
 board_build.f_cpu         = 80000000L
-build_flags               = ${debug_flags.build_flags} ${mqtt_flags.build_flags}
+build_flags               = ${debug_flags.build_flags} ${mqtt_flags.build_flags} ${minimal_size.build_flags}
 build_unflags             = -DDEBUG_ESP_PORT
 lib_deps                  = 
 lib_ignore                = ESP32_ping, ESP32WebServer, IRremoteESP8266, HeatpumpIR
@@ -754,7 +764,7 @@ build_unflags             = ${esp8266_4M.build_unflags}
 
 [env:test_core_260_sdk222_alpha_ESP8266_4M]
 platform                  = ${testing_beta.platform}
-build_flags               = ${esp8266_4M.build_flags} ${testing_beta.build_flags} -DPIO_FRAMEWORK_ARDUINO_ESPRESSIF_SDK22x
+build_flags               = ${esp8266_4M.build_flags} ${testing_beta.build_flags} -DPIO_FRAMEWORK_ARDUINO_ESPRESSIF_SDK22y
 board                     = ${esp8266_4M.board}
 board_build.f_cpu         = ${esp8266_4M.board_build.f_cpu}
 board_build.flash_mode    = ${esp8266_4M.board_build.flash_mode}
@@ -784,7 +794,7 @@ build_unflags             = ${esp8266_16M.build_unflags}
 
 [env:normal_core_260_sdk222_alpha_ESP8266_16M]
 platform                  = ${testing_beta.platform}
-build_flags               = ${esp8266_16M.build_flags} ${normal_beta.build_flags} -DPIO_FRAMEWORK_ARDUINO_ESPRESSIF_SDK22x
+build_flags               = ${esp8266_16M.build_flags} ${normal_beta.build_flags} -DPIO_FRAMEWORK_ARDUINO_ESPRESSIF_SDK22y
 board                     = ${esp8266_16M.board}
 board_build.f_cpu         = ${esp8266_16M.board_build.f_cpu}
 board_build.flash_mode    = ${esp8266_16M.board_build.flash_mode}
@@ -795,7 +805,7 @@ build_unflags             = ${esp8266_16M.build_unflags}
 
 [env:test_core_260_sdk222_alpha_ESP8266_16M]
 platform                  = ${testing_beta.platform}
-build_flags               = ${esp8266_16M.build_flags} ${testing_beta.build_flags} -DPIO_FRAMEWORK_ARDUINO_ESPRESSIF_SDK22x
+build_flags               = ${esp8266_16M.build_flags} ${testing_beta.build_flags} -DPIO_FRAMEWORK_ARDUINO_ESPRESSIF_SDK22y
 board                     = ${esp8266_16M.board}
 board_build.f_cpu         = ${esp8266_16M.board_build.f_cpu}
 board_build.flash_mode    = ${esp8266_16M.board_build.flash_mode}
