+++ pre_extra_script.py
@@ -12,7 +12,7 @@ Import("env")
 # - frameworks
 # - dependent libraries
 env.Append(CPPDEFINES=[
-  "PIO_FRAMEWORK_ARDUINO_ESPRESSIF_SDK22x",
+  "PIO_FRAMEWORK_ARDUINO_ESPRESSIF_SDK22y",
   "CONTROLLER_SET_ALL",
   "NOTIFIER_SET_NONE",
   "PLUGIN_SET_ONLY_SWITCH",
