+++ src/ESPEasy-Globals.h
@@ -498,28 +498,6 @@ void check_size() {
 #define ZERO_FILL(S)  memset((S), 0, sizeof(S))
 #define ZERO_TERMINATE(S)  S[sizeof(S) - 1] = 0
 
-// Forward declaration
-struct ControllerSettingsStruct;
-static String getUnknownString();
-void scheduleNextDelayQueue(unsigned long id, unsigned long nextTime);
-String LoadControllerSettings(int ControllerIndex, ControllerSettingsStruct& controller_settings);
-String get_formatted_Controller_number(int controller_index);
-bool loglevelActiveFor(byte logLevel);
-void addToLog(byte loglevel, const String& string);
-void addToLog(byte logLevel, const __FlashStringHelper* flashString);
-void statusLED(boolean traffic);
-void backgroundtasks();
-uint32_t getCurrentFreeStack();
-uint32_t getFreeStackWatermark();
-bool canYield();
-
-bool getBitFromUL(uint32_t number, byte bitnr);
-void setBitToUL(uint32_t& number, byte bitnr, bool value);
-
-void serialHelper_getGpioNames(struct EventStruct *event, bool rxOptional=false, bool txOptional=false);
-
-fs::File tryOpenFile(const String& fname, const String& mode);
-
 enum SettingsType {
   BasicSettings_Type = 0,
   TaskSettings_Type,
@@ -552,6 +530,8 @@ bool showSettingsFileLayout = false;
 #include "ESPEasyTimeTypes.h"
 #include "StringProviderTypes.h"
 #include "ESPeasySerial.h"
+#include "ESPEasy_fdwdecl.h"
+#include "WebServer_fwddecl.h"
 #include "I2CTypes.h"
 #include <I2Cdev.h>
 #include <map>
@@ -709,10 +689,6 @@ bool P037_MQTTImport_connected = false;
 // udp protocol stuff (syslog, global sync, node info list, ntp time)
 WiFiUDP portUDP;
 
-bool resolveHostByName(const char* aHostname, IPAddress& aResult);
-bool connectClient(WiFiClient& client, const char* hostname, uint16_t port);
-bool connectClient(WiFiClient& client, IPAddress ip, uint16_t port);
-
 class TimingStats;
 
 /*********************************************************************************************\
@@ -728,27 +704,6 @@ struct CRCStruct{
   uint32_t numberOfCRCBytes=0;
 }CRCValues;
 
-
-bool WiFiConnected(uint32_t timeout_ms);
-bool WiFiConnected();
-bool hostReachable(const IPAddress& ip);
-bool hostReachable(const String& hostname);
-void formatMAC(const uint8_t* mac, char (&strMAC)[20]);
-String to_json_object_value(const String& object, const String& value);
-
-
-bool I2C_read_bytes(uint8_t i2caddr, I2Cdata_bytes& data);
-bool I2C_write8_reg(uint8_t i2caddr, byte reg, byte value);
-uint8_t I2C_read8_reg(uint8_t i2caddr, byte reg, bool * is_ok = NULL);
-uint16_t I2C_read16_reg(uint8_t i2caddr, byte reg);
-int32_t I2C_read24_reg(uint8_t i2caddr, byte reg);
-uint16_t I2C_read16_LE_reg(uint8_t i2caddr, byte reg);
-int16_t I2C_readS16_reg(uint8_t i2caddr, byte reg);
-int16_t I2C_readS16_LE_reg(uint8_t i2caddr, byte reg);
-I2Cdev i2cdev;
-
-bool safe_strncpy(char* dest, const String& source, size_t max_size);
-bool safe_strncpy(char* dest, const char* source, size_t max_size);
 
 
 
@@ -2288,6 +2243,9 @@ enum DeviceModel {
   DeviceModel_MAX
 };
 
+bool modelMatchingFlashSize(DeviceModel model, int size_MB);
+
+
 struct ResetFactoryDefaultPreference_struct {
   ResetFactoryDefaultPreference_struct(uint32_t preference = 0) : _preference(preference) {}
 
@@ -2425,7 +2383,6 @@ struct GpioFactorySettingsStruct {
   int8_t i2c_scl = DEFAULT_PIN_I2C_SCL;
 };
 
-bool modelMatchingFlashSize(DeviceModel model, int size_MB);
 void addPredefinedPlugins(const GpioFactorySettingsStruct& gpio_settings);
 void addPredefinedRules(const GpioFactorySettingsStruct& gpio_settings);
 
@@ -2613,5 +2570,7 @@ String getPackedFromPlugin(struct EventStruct *event, uint8_t sampleSetCount);
 #include "ESPEasyWiFiEvent.h"
 #define SPIFFS_CHECK(result, fname) if (!(result)) { return(FileError(__LINE__, fname)); }
 #include "WebServer_Rules.h"
+
+
 
 #endif /* ESPEASY_GLOBALS_H_ */
