+++ src/StringProviderTypes.h
@@ -99,12 +99,6 @@ String getInternalLabel(LabelType::Enum label);
 String getLabel(LabelType::Enum label);
 String getValue(LabelType::Enum label);
 String getExtendedValue(LabelType::Enum label);
-void stream_next_json_object_value(LabelType::Enum label);
-void stream_last_json_object_value(LabelType::Enum label);
-void addRowLabelValue(LabelType::Enum label);
-void addRowLabelValue_copy(LabelType::Enum label);
-void addFormCheckBox(LabelType::Enum label, boolean checked, bool disabled = false);
-void addFormCheckBox_disabled(LabelType::Enum label, boolean checked);
 
 
 struct FileType {
